+++ src/Tests/dotnet-build.Tests/GivenDotnetBuildBuildsCsproj.cs
@@ -2,15 +2,17 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Generic;
 using System.IO;
-using FluentAssertions;
-using Xunit;
 using System.Linq;
+using System.Xml.Linq;
+using FluentAssertions;
 using Microsoft.NET.TestFramework;
 using Microsoft.NET.TestFramework.Assertions;
 using Microsoft.NET.TestFramework.Commands;
+using Microsoft.NET.TestFramework.ProjectConstruction;
+using Xunit;
 using Xunit.Abstractions;
-using Microsoft.DotNet.Tools;
 
 namespace Microsoft.DotNet.Cli.Build.Tests
 {
@@ -222,6 +224,89 @@ namespace Microsoft.DotNet.Cli.Build.Tests
                .Pass()
                .And
                .NotHaveStdOutContaining("NETSDK1179");
+        }
+
+        [Theory]
+        [InlineData("roslyn3.9")]
+        [InlineData("roslyn4.0")]
+        public void It_resolves_analyzers_tareting_mulitple_roslyn_versions(string compilerApiVersion)
+        {
+            var testProject = new TestProject()
+            {
+                TargetFrameworks = "netstandard2.0"
+            };
+
+            //  Disable analyzers built in to the SDK so we can more easily test the ones coming from NuGet packages
+            testProject.AdditionalProperties["EnableNETAnalyzers"] = "false";
+
+            testProject.ProjectChanges.Add(project =>
+            {
+                var propertyGroup = XElement.Parse($@"
+  <PropertyGroup>
+    <CompilerApiVersion>{compilerApiVersion}</CompilerApiVersion>
+  </PropertyGroup>");
+
+                project.Root.Add(propertyGroup);
+
+                var itemGroup = XElement.Parse(@"
+  <ItemGroup>
+    <PackageReference Include=""Library.ContainsAnalyzer"" Version=""1.0.0"" />
+    <PackageReference Include=""Library.ContainsAnalyzer2"" Version=""1.0.0"" />
+  </ItemGroup>");
+
+                project.Root.Add(itemGroup);
+            });
+
+            var testAsset = _testAssetsManager.CreateTestProject(testProject, identifier: compilerApiVersion);
+
+            NuGetConfigWriter.Write(testAsset.Path, TestContext.Current.TestPackages);
+
+            var command = new GetValuesCommand(testAsset,
+                "Analyzer",
+                GetValuesCommand.ValueType.Item);
+
+            command.Execute().Should().Pass();
+
+            var analyzers = command.GetValues();
+
+            switch (compilerApiVersion)
+            {
+                case "roslyn3.9":
+                    analyzers.Select(RelativeNuGetPath).Should().BeEquivalentTo(
+                        "library.containsanalyzer/1.0.0/analyzers/dotnet/roslyn3.9/cs/Library.ContainsAnalyzer.dll",
+                        "library.containsanalyzer2/1.0.0/analyzers/dotnet/roslyn3.8/cs/Library.ContainsAnalyzer2.dll"
+                        );
+                    break;
+
+                case "roslyn4.0":
+                    analyzers.Select(RelativeNuGetPath).Should().BeEquivalentTo(
+                        "library.containsanalyzer/1.0.0/analyzers/dotnet/roslyn4.0/cs/Library.ContainsAnalyzer.dll",
+                        "library.containsanalyzer2/1.0.0/analyzers/dotnet/roslyn3.10/cs/Library.ContainsAnalyzer2.dll"
+                        );
+                    break;
+
+                default:
+                    throw new ArgumentOutOfRangeException(nameof(compilerApiVersion));
+            }
+        }
+
+        static readonly List<string> nugetRoots = new List<string>()
+        {
+            TestContext.Current.NuGetCachePath,
+            Path.Combine(FileConstants.UserProfileFolder, ".dotnet", "NuGetFallbackFolder")
+        };
+
+        static string RelativeNuGetPath(string absoluteNuGetPath)
+        {
+            foreach (var nugetRoot in nugetRoots)
+            {
+                if (absoluteNuGetPath.StartsWith(nugetRoot + Path.DirectorySeparatorChar))
+                {
+                    return absoluteNuGetPath.Substring(nugetRoot.Length + 1)
+                                .Replace(Path.DirectorySeparatorChar, '/');
+                }
+            }
+            throw new InvalidDataException("Expected path to be under a NuGet root: " + absoluteNuGetPath);
         }
     }
 }
